Revert "Start tracking edits performed by the agent (#27064)"
This reverts commit ac5dafc6b28ae32960d7692b7d79f95c27b01f37.
