+++ crates/assistant2/src/thread.rs
@@ -1,6 +1,5 @@
 use std::fmt::Write as _;
 use std::io::Write;
-use std::ops::Range;
 use std::sync::Arc;
 
 use anyhow::{Context as _, Result};
@@ -976,10 +975,6 @@ impl Thread {
         })
     }
 
-    pub fn project(&self) -> &Entity<Project> {
-        &self.project
-    }
-
     /// Create a snapshot of the current project state including git information and unsaved buffers.
     fn project_snapshot(
         project: Entity<Project>,
@@ -1126,18 +1121,6 @@ impl Thread {
         }
 
         Ok(String::from_utf8_lossy(&markdown).to_string())
-    }
-
-    pub fn review_edits_in_range(
-        &mut self,
-        buffer: Entity<language::Buffer>,
-        buffer_range: Range<language::Anchor>,
-        accept: bool,
-        cx: &mut Context<Self>,
-    ) -> Task<Result<()>> {
-        self.action_log.update(cx, |action_log, cx| {
-            action_log.review_edits_in_range(buffer, buffer_range, accept, cx)
-        })
     }
 
     pub fn action_log(&self) -> &Entity<ActionLog> {
