+++ crates/assistant_tool/Cargo.toml
@@ -13,7 +13,6 @@ path = "src/assistant_tool.rs"
 
 [dependencies]
 anyhow.workspace = true
-buffer_diff.workspace = true
 collections.workspace = true
 clock.workspace = true
 derive_more.workspace = true
@@ -24,12 +23,3 @@ parking_lot.workspace = true
 project.workspace = true
 serde.workspace = true
 serde_json.workspace = true
-
-[dev-dependencies]
-buffer_diff = { workspace = true, features = ["test-support"] }
-collections = { workspace = true, features = ["test-support"] }
-clock = { workspace = true, features = ["test-support"] }
-gpui = { workspace = true, features = ["test-support"] }
-language = { workspace = true, features = ["test-support"] }
-language_model = { workspace = true, features = ["test-support"] }
-project = { workspace = true, features = ["test-support"] }
